+++ tests/test_httpx.py
@@ -1,18 +1,19 @@
 import datetime
 
-import aiohttp
+import httpx
+
 import pytest
 
-from .. import aiohttp as gh_aiohttp
-from .. import sansio
+from gidgethub import httpx as gh_httpx
+from gidgethub import sansio
 
 
 @pytest.mark.asyncio
 async def test_sleep():
     delay = 1
     start = datetime.datetime.now()
-    async with aiohttp.ClientSession() as session:
-        gh = gh_aiohttp.GitHubAPI(session, "gidgethub")
+    async with httpx.AsyncClient() as client:
+        gh = gh_httpx.GitHubAPI(client, "gidgethub")
         await gh.sleep(delay)
     stop = datetime.datetime.now()
     assert (stop - start) > datetime.timedelta(seconds=delay)
@@ -22,8 +23,8 @@ async def test_sleep():
 async def test__request():
     """Make sure that that abstract method is implemented properly."""
     request_headers = sansio.create_headers("gidgethub")
-    async with aiohttp.ClientSession() as session:
-        gh = gh_aiohttp.GitHubAPI(session, "gidgethub")
+    async with httpx.AsyncClient() as client:
+        gh = gh_httpx.GitHubAPI(client, "gidgethub")
         aio_call = await gh._request(
             "GET", "https://api.github.com/rate_limit", request_headers
         )
@@ -34,7 +35,7 @@ async def test__request():
 @pytest.mark.asyncio
 async def test_get():
     """Integration test."""
-    async with aiohttp.ClientSession() as session:
-        gh = gh_aiohttp.GitHubAPI(session, "gidgethub")
+    async with httpx.AsyncClient() as client:
+        gh = gh_httpx.GitHubAPI(client, "gidgethub")
         data = await gh.getitem("/rate_limit")
     assert "rate" in data